Исправление замечаний по код-ревью
