+++ src/test/java/FelineTest.java
@@ -3,10 +3,11 @@ import org.junit.Test;
 import java.util.List;
 import static org.junit.Assert.assertEquals;
 public class FelineTest {
-    Feline feline = new Feline();
+
 
     @Test
     public void eatMeatReturnReturnCorrectResult() throws Exception{
+        Feline feline = new Feline();
         List<String> expectedResult = List.of("Животные", "Птицы", "Рыба");
         List<String> actualResult = feline.eatMeat();
         assertEquals (expectedResult, actualResult);
@@ -14,6 +15,7 @@ public class FelineTest {
 
     @Test
     public void getFamilyReturnCorrectResult(){
+        Feline feline = new Feline();
         String expectedResult = "Кошачьи";
         String actualResult = feline.getFamily();
         assertEquals (expectedResult, actualResult);
@@ -21,6 +23,7 @@ public class FelineTest {
 
     @Test
     public void getKittensWithOutParameterReturnCorrectResult(){
+        Feline feline = new Feline();
         int expectedResult = 1;
         int actualResult = feline.getKittens();
         assertEquals (expectedResult, actualResult);
@@ -28,6 +31,7 @@ public class FelineTest {
 
     @Test
     public void getKittensWithParameterReturnCorrectResult(){
+        Feline feline = new Feline();
         int expectedResult = 5;
         int actualResult = feline.getKittens(5);
         assertEquals (expectedResult, actualResult);
