+++ src/test/java/ParameterizedLionTest.java
@@ -1,0 +1,37 @@
+import com.example.Feline;
+import com.example.Lion;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.mockito.Mock;
+
+import static org.junit.Assert.assertEquals;
+
+@RunWith(Parameterized.class)
+public class ParameterizedLionTest {
+    @Mock
+    Feline feline;
+    private final String sex;
+    private final boolean hasMane;
+
+    public ParameterizedLionTest(String sex, boolean hasMane) {
+        this.sex = sex;
+        this.hasMane = hasMane;
+    }
+
+    @Parameterized.Parameters
+    public static Object[][] getData() {
+        return new Object[][]{
+                {"Самец", true},
+                {"Самка", false},
+                {"Неопределенно", false},
+        };
+    }
+
+    @Test
+    public void doesHaveManeReturnCorrectBoolResult() throws Exception {
+        Lion lion = new Lion(sex, feline);
+        boolean actualResult = lion.doesHaveMane();
+        assertEquals(hasMane, actualResult);
+    }
+}
